Crate polkadot_node_core_dispute_coordinator

Implements the dispute coordinator subsystem.

This is the central subsystem of the node-side components which participate in disputes. This subsystem wraps a database which tracks all statements observed by all validators over some window of sessions. Votes older than this session window are pruned.

This subsystem will be the point which produce dispute votes, either positive or negative, based on locally-observed validation results as well as a sink for votes received by other subsystems. When importing a dispute vote from another node, this will trigger the dispute participation subsystem to recover and validate the block and call back to this subsystem.

Structs

Config

Configuration for the dispute coordinator subsystem.

Enums

DisputeCoordinatorSubsystem

The disputes coordinator subsystem, abstracts dummy and real implementations.